Update SortTestClient.java
Two changes for this, one that is definitely needed as it restores the intended aspects of the test client and the other is a change for readability. The change from the one sort implementation to two change the values of how the values work from doubling each time (2 to 4, 4 to 8, 8 to 16, etc.) to some thing where the initial values is multiplied to the number of the initial value to the amount of cycles to the power of 2 (2 to 8, 8 to 18, 18 to 32, etc.), so  change it to the first run is just 2000, and there rest is multiplied by 2 each time. The other change is that  as the printouts listed each type at the same time, it would be difficult to find patterns in the results of the same type without searching and listing each type individually (i.e. find and copy-paste each ascending value, find and copy-paste each dup ascending value, etc.) so changed it to where the code finishes the type that it is on before moving to the next. Finally, simple matter of "StopWatch" (check other commit notes for reason for quotations)  not being used in the client by itself, so the initialization for it was removed.